app/requirements.txt
requests
PyGithub
python-dotenv
uvicorn
sseclient-py
fastapi
PyYAML
